create file handlers POST endpoint
